Split automation timeout enforcement
